refactor: dedupe unrefSleepClock and the 60-char preview-truncation idiom
Two small, verified duplication clusters from a codebase-wide refactoring
audit:

- unrefSleepClock: an identical 22-line Effect Clock wrapper (unref'd
  setTimeout sleep) was hand-copied in UsageLogService.ts and
  PollingSourceBase.ts. Moved to a single shared
  src/utils/system/unrefSleepClock.ts.

- previewLabel: truncateWithEllipsis(text, 60) was repeated at 8 call
  sites across 7 tool files as the shared '60-char preview' convention.
  Added a previewLabel() helper (stringUtils.ts) wrapping the constant,
  and switched all 8 sites to it.

Both are pure consolidations with no behavior change. Updated the
browser-safe-utils module counts in CLAUDE.md/AGENTS.md for the new file.

Verified: full workspace typecheck, full workspace lint, all directly
affected test suites (152 tests), dead-code ratchet, browser-safe-utils
drift guard, and prettier — all clean.

Investigated but intentionally left alone (documented in the PR
description): the desktop/extension host-request routers, the
codex/xai OAuth provider families, and the CLI's initCliPlatform
bootstrap line — each looked like duplication in a codebase-wide scan
but turned out on close reading to either diverge in ways that don't
mechanically consolidate, or touch security-sensitive/unverifiable
surfaces (Electron main-process wiring, OAuth flows) this session has
no way to runtime-verify.
