+++ CLAUDE.md
@@ -59,7 +59,7 @@ Things the tree won't tell you:
 - **`src/utils/` is host-agnostic, not universally browser-safe.** Exactly five
   modules are browser-reachable today: `@utils/core`,
   `@utils/core/keyedMutex`, `@utils/errors/errorMessage`,
-  `@utils/files/pastedImageName`, and `@utils/text/stringUtils`. The other 59
+  `@utils/files/pastedImageName`, and `@utils/text/stringUtils`. The other 60
   TypeScript modules are not browser-reachable and must not be assumed
   browser-safe. Side-specific helpers still belong in `frontend/` or `common/`.
   (`scripts/check-browser-safe-utils.mjs` enforces the count and reachable set.)
