+++ src/tools/agentCliShared.ts
@@ -46,7 +46,7 @@ import {
 import { executed } from '@tools/core/result';
 import { generateExecutionId } from '@utils/core';
 import { toErrorMessage } from '@utils/errors/errorMessage';
-import { truncateWithEllipsis } from '@utils/text/stringUtils';
+import { previewLabel } from '@utils/text/stringUtils';
 
 import {
   childStreamDescription,
@@ -133,7 +133,7 @@ async function queueAgentCliFollowUp(
     );
   }
 
-  const preview = truncateWithEllipsis(prompt, 60);
+  const preview = previewLabel(prompt);
   // A queued follow-up whose wake failed is still queued: it is delivered
   // when the agent is resumed, so the caller must not offer it again.
   const wakeFailed = result.status === 'queued' && result.wake === 'failed';
