+++ src/tools/claudeAgent.ts
@@ -53,7 +53,7 @@ import {
   isNonEmptyString,
   linkAbortSignals,
 } from '@utils/core';
-import { truncateWithEllipsis } from '@utils/text/stringUtils';
+import { previewLabel } from '@utils/text/stringUtils';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 
 // Local file imports
@@ -573,7 +573,7 @@ async function launchClaudeAgentSession(
   const env = await config.buildClaudeAgentEnv();
   const pathToClaudeCodeExecutable = await findClaudeBinaryPath();
   const agentConfig = config.buildClaudeAgentConfig(input.prompt);
-  const preview = truncateWithEllipsis(input.prompt, 60);
+  const preview = previewLabel(input.prompt);
 
   return launchAgentCliSession({
     parentStreamId,
